Resolve "Add production compile flags"

Closes #47

See merge request Kickness/cpm!24

+++ src/utils/render.h
@@ -3,7 +3,7 @@
 
 #include <stddef.h>
 
-char* render(const char* template, const char** args, unsigned long);
-size_t render_length(const char* template, const char** args, unsigned long);
+char* render(const char* template, const char** args, int);
+long render_length(const char* template, const char** args, int);
 
 #endif
